refactor login elements out of splashRoot.js
